refactor topic address map parsing into the kafkarestclientfactory
