+++ lib/kafka_rest_client.js
@@ -126,25 +126,6 @@ KafkaRestClient.prototype.getTopicRequestBody = function getTopicRequestBody(pro
     });
 };
 
-KafkaRestClient.prototype.getTopicToUrlMapping = function getTopicToUrlMapping(body) {
-    var self = this;
-    try {
-        var urlToTopicsJson = JSON.parse(body);
-        var topicToUrlMapping = {};
-        var urlArray = Object.keys(urlToTopicsJson);
-        for (var i = 0; i < urlArray.length; i++) {
-            var reqUrl = urlArray[i];
-            var topicsArray = urlToTopicsJson[reqUrl];
-            for (var j = 0; j < topicsArray.length; j++) {
-                topicToUrlMapping[topicsArray[j]] = reqUrl;
-            }
-        }
-        return topicToUrlMapping;
-    } catch(e) {
-        return self.cachedTopicToUrlMapping;
-    }
-};
-
 /*
  ProduceMessage {
    string topic,
