+++ lib/kafka_rest_client_factory.js
@@ -19,6 +19,7 @@
 // THE SOFTWARE.
 
 'use strict';
+var KafkaRestClient = require('./kafka_rest_client');
 
 var instance = null;
 
@@ -31,12 +32,12 @@ function KafkaRestClientFactory(options, callback) {
     self.proxyPort = options.proxyPort;
 
     # map topic to address
-    self.topic_address_map = {};
+    self.topicAddressMap = {};
 
     # map one KafkaRestClient per address
-    self.address_client_map = {};
+    self.addressClientMap = {};
 
-    self.loadTopicMap();
+    self.loadTopicMap(callback);
 };
 
 KafkaRestClientFactory.prototype.getKafkaRestClientFactory = function getKafkaRestClientFactory(options) {
@@ -47,8 +48,39 @@ KafkaRestClientFactory.prototype.getKafkaRestClientFactory = function getKafkaRe
     return instance;
 };
 
-KafkaRestClientFactory.prototype.loadTopicMap = function loadTopicMap() {
-    var url = util.format('http://[%s]:[%s]/topics', self.proxyHost, self.proxyPort);
+function formatTopicAddressMapJSON(body) {
+    var self = this;
+    try {
+        var urlToTopicsJson = JSON.parse(body);
+        var topicToUrlMapping = {};
+        var urlArray = Object.keys(urlToTopicsJson);
+        for (var i = 0; i < urlArray.length; i++) {
+            var reqUrl = urlArray[i];
+            var topicsArray = urlToTopicsJson[reqUrl];
+            for (var j = 0; j < topicsArray.length; j++) {
+                topicToUrlMapping[topicsArray[j]] = reqUrl;
+            }
+        }
+
+        return topicToUrlMapping;
+    } catch(e) {
+        return self.topicAddressMap;
+    }
+};
+
+KafkaRestClientFactory.prototype.loadTopicMap = function loadTopicMap(callback) {
+    KafkaRestClient.getAllTopicsRequest(self.proxyHost, self.proxyPort, function updateTopicAddressMap(err, res) {
+        if (err) {
+            callback(err);
+        } else {
+            var topicAddressMap = formatTopicAddressMapJSON(res);
+
+            if(Object.keys(topicAddressMap).length > 0) {
+                self.topicAddressMap = formatTopicAddressMapJSON(res);
+                //todo makes Address Client Map
+            }
+        }
+    });
 };
 
 module.exports = KafkaRestClientFactory;
